chore(cmd): migrate stats command to cobra style

# Conflicts:
#	cmd/bbolt/command_root.go
